rename and moved route config file
