+++ src/config/index.js
@@ -1,1 +1,2 @@
 export { default as Roles } from './Roles';
+export { default as PrivateRoutesConfig } from './PrivateRoutesConfig';
